requirements.txt
beautifulsoup4>=4.13.0
Flask>=3.1.0
numpy>=1.25.0
openai>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.32.0
streamlit>=1.25.0
